+++ source/rule.cpp
@@ -1,13 +1,13 @@
 /* rule.h */
 #include <rule.h>
 
-std::vector<double> rule::lambda;
+std::vector<float> rule::lambda;
 
 void rule::evaluate_score()
 {
     unsigned int size = target_rule_body.size();
-    const double penalty = -0.4342944819032518276511289189166;
-    double terminal_num_weight = lambda[4];
+    const float penalty = -0.4342944819032518276511289189166;
+    float terminal_num_weight = lambda[4];
 
     for (unsigned int i = 0; i < size; i++) {
         const symbol* sym = target_rule_body[i];
@@ -19,7 +19,7 @@ void rule::evaluate_score()
 
     for (unsigned int i = 0; i < size; i++) {
         const feature& fea = feature_set[i];
-        double weight = lambda[i];
+        float weight = lambda[i];
 
         score += weight * fea.get_score();
     }
